# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PEPL API",
    "version": "1.0.0",
    "description": "Multi-tenant HR and employee-operations API.\n\n## Conventions\n\n- **Tenancy is implicit.** The company is resolved from the session token. There is no\n  tenant id in any path, header or body, and supplying one would have no effect.\n- **404, not 403, for scope.** A record belonging to another employee outside the caller's\n  data scope returns 404. Confirming that a record exists is itself a disclosure.\n- **Stable error codes.** Every error carries `error.code` — switch on that, not on the\n  status or the message. `error.requestId` correlates with the server log.\n- **Reasons.** Any change that moves money or rewrites history requires a `reason`, and\n  the API refuses the request without one.\n\n## Building a UI against this\n\nCall `GET /api/v1/me` first. It returns the caller's `permissions`, data `scope` and the\nset of enabled `modules`. Render navigation from that rather than from the role name —\na tenant can define custom roles, so role names are not a reliable switch.\n\nA module the tenant has disabled, or has not purchased, returns 403 `MODULE_NOT_AVAILABLE`.\nHide it rather than showing a disabled teaser."
  },
  "servers": [
    {
      "url": "http://localhost:4010",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/v1/attendance/punch": {
      "post": {
        "summary": "Record a punch (idempotent by clientPunchId)",
        "tags": [
          "attendance"
        ],
        "operationId": "post_api_v1_attendance_punch",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied, or the module is not enabled for this company",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found — also returned for a record outside the caller's data scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state (a frozen period, a locked run)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed, or a domain rule refused the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Requires the `attendance.read` permission.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "direction": "in",
                "localDate": "2026-09-14",
                "clientPunchId": "offline-1",
                "geo": {
                  "lat": 17.4,
                  "lng": 78.4
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/attendance": {
      "get": {
        "summary": "Daily attendance for a date range",
        "tags": [
          "attendance"
        ],
        "operationId": "get_api_v1_attendance",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied, or the module is not enabled for this company",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found — also returned for a record outside the caller's data scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state (a frozen period, a locked run)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed, or a domain rule refused the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Requires the `attendance.read` permission."
      }
    },
    "/api/v1/attendance/corrections": {
      "post": {
        "summary": "Correct a day. Instant while the period is open; never rewrites a frozen one",
        "tags": [
          "attendance"
        ],
        "operationId": "post_api_v1_attendance_corrections",
        "parameters": [],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied, or the module is not enabled for this company",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found — also returned for a record outside the caller's data scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state (a frozen period, a locked run)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed, or a domain rule refused the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Requires the `attendance.correct` permission.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "employeeId": "…",
                "workDate": "2026-09-14",
                "action": "mark_present",
                "reason": "biometric device offline"
              }
            }
          }
        }
      }
    },
    "/api/v1/attendance/periods/{id}/status": {
      "post": {
        "summary": "Close, reopen or freeze an attendance period",
        "tags": [
          "attendance"
        ],
        "operationId": "post_api_v1_attendance_periods__id__status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied, or the module is not enabled for this company",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found — also returned for a record outside the caller's data scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflicts with the current state (a frozen period, a locked run)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed, or a domain rule refused the request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Requires the `attendance.close_period` permission.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "status": "closed"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "The token from POST /api/v1/auth/login"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Stable machine code. Switch on this."
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object"
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        },
        "example": {
          "error": {
            "code": "CONFIG_EFFECTIVE_DATE_REQUIRED",
            "message": "\"payroll.lop_basis\" affects payroll and must carry an effective date, so a locked run stays reproducible",
            "requestId": "…"
          }
        }
      }
    }
  },
...
}